new files added

+++ Getri_FinalProject_MVC_API/ViewModel/ProductCreateViewModel.cs
@@ -10,7 +10,7 @@ namespace Getri_FinalProject_MVC_API.ViewModel
 
 		public string? ProductDescription { get; set; }
 
-		public decimal? ProductPrice { get; set; }
+		public int? ProductPrice { get; set; }
 
 		public int? CategoryId { get; set; }
